feat: hide more of the partiality-handling code
